requirements.txt
# AI-Driven DeFi Yield Rebalancing System
# Python Dependencies

# ============================================================================
# Core Dependencies
# ============================================================================
python-dotenv==1.0.0
pydantic==2.5.3
pyyaml==6.0.1

# ============================================================================
# Deep Learning & ML
# ============================================================================
torch>=2.2.0
torchvision>=0.17.0
pytorch-lightning>=2.2.0
tensorboard>=2.16.0

# Reinforcement Learning
stable-baselines3[extra]>=2.2.1
gymnasium==0.29.1
sb3-contrib==2.2.1

# Traditional ML
xgboost==2.0.3
lightgbm==4.1.0
scikit-learn==1.3.2
shap==0.44.0

# Model Optimization
optuna==3.5.0
hyperopt==0.2.7

# ============================================================================
# Data Processing & Analysis
# ============================================================================
numpy==1.26.2
pandas==2.1.4
polars==0.19.19  # Faster alternative to pandas
pyarrow==14.0.1

# Feature Engineering
ta==0.11.0  # Technical analysis library
statsmodels==0.14.1

# ============================================================================
# Blockchain & Web3
# ============================================================================
web3==6.13.0
eth-account==0.10.0
eth-abi==4.2.1
eth-utils==2.3.1

# The Graph Integration
gql==3.4.1
requests-toolbelt==1.0.0

# ============================================================================
# Database
# ============================================================================
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1  # Database migrations

# TimescaleDB (requires PostgreSQL with TimescaleDB extension)
timescaledb>=0.0.4

# ============================================================================
# Caching & Message Queue
# ============================================================================
redis==5.0.1
celery==5.3.4

# ============================================================================
# API & Web Framework
# ============================================================================
fastapi==0.109.0
uvicorn[standard]==0.26.0
httpx==0.26.0
websockets==12.0

# ============================================================================
# Monitoring & Logging
# ============================================================================
prometheus-client==0.19.0
sentry-sdk==1.39.1
structlog==24.1.0
python-json-logger==2.0.7

# ============================================================================
# Testing
# ============================================================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
hypothesis==6.92.2  # Property-based testing

# ============================================================================
# Code Quality
# ============================================================================
black==23.12.1
ruff==0.1.9
mypy==1.8.0
pre-commit==3.6.0

# ============================================================================
# Utilities
# ============================================================================
click==8.1.7  # CLI framework
rich==13.7.0  # Terminal formatting
tqdm==4.66.1  # Progress bars
schedule==1.2.0  # Job scheduling

# ============================================================================
# Security
# ============================================================================
cryptography==41.0.7
pycryptodome==3.19.0

# ============================================================================
# Data Visualization (for analysis)
# ============================================================================
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# ============================================================================
# Jupyter (for experimentation)
# ============================================================================
jupyter==1.0.0
ipywidgets==8.1.1
notebook==7.0.6

# ============================================================================
# External API Clients
# ============================================================================
dune-client>=1.3.0  # Dune Analytics
pycoingecko>=3.0.0  # CoinGecko API
